+++ src/transport/getblock.rs
@@ -10,7 +10,7 @@
 /// ```rust
 /// use bitcoin_rpc_codegen::client::latest::getblock;
 ///
-/// let client = Client::new("http://127.0.0.1:8332", auth);
+/// let client = Client::new("http://127.0.0.1:18443", auth);
 /// let result = client.getblock(/* params */).await?;
 /// ```
 
@@ -26,13 +26,13 @@ pub struct GetblockResponse {
         #[serde(skip_serializing_if = "Option::is_none")]
     pub hash: Option<String>,
         #[serde(skip_serializing_if = "Option::is_none")]
-    pub confirmations: Option<f64>,
+    pub confirmations: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
     pub size: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
     pub strippedsize: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
-    pub weight: Option<f64>,
+    pub weight: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
     pub height: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
@@ -48,7 +48,7 @@ pub struct GetblockResponse {
         #[serde(skip_serializing_if = "Option::is_none")]
     pub mediantime: Option<serde_json::Value>,
         #[serde(skip_serializing_if = "Option::is_none")]
-    pub nonce: Option<f64>,
+    pub nonce: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
     pub bits: Option<String>,
         #[serde(skip_serializing_if = "Option::is_none")]
@@ -56,7 +56,7 @@ pub struct GetblockResponse {
         #[serde(skip_serializing_if = "Option::is_none")]
     pub chainwork: Option<String>,
         #[serde(skip_serializing_if = "Option::is_none")]
-    pub n_tx: Option<f64>,
+    pub n_tx: Option<u64>,
         #[serde(skip_serializing_if = "Option::is_none")]
     pub previousblockhash: Option<bitcoin::BlockHash>,
         #[serde(skip_serializing_if = "Option::is_none")]
