+++ src/transport/getblockcount.rs
@@ -8,7 +8,7 @@
 /// ```rust
 /// use bitcoin_rpc_codegen::client::latest::getblockcount;
 ///
-/// let client = Client::new("http://127.0.0.1:8332", auth);
+/// let client = Client::new("http://127.0.0.1:18443", auth);
 /// let result = client.getblockcount().await?;
 /// ```
 
@@ -19,7 +19,7 @@ use transport::{TransportTrait, TransportError};
     /// The genesis block has height 0.
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(transparent)]
-pub struct GetblockcountResponse(pub f64);
+pub struct GetblockcountResponse(pub u64);
