+++ src/transport/getconnectioncount.rs
@@ -7,7 +7,7 @@
 /// ```rust
 /// use bitcoin_rpc_codegen::client::latest::getconnectioncount;
 ///
-/// let client = Client::new("http://127.0.0.1:8332", auth);
+/// let client = Client::new("http://127.0.0.1:18443", auth);
 /// let result = client.getconnectioncount().await?;
 /// ```
 
@@ -17,7 +17,7 @@ use transport::{TransportTrait, TransportError};
 /// Returns the number of connections to other nodes.
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(transparent)]
-pub struct GetconnectioncountResponse(pub f64);
+pub struct GetconnectioncountResponse(pub u64);
