+++ src/transport/getmininginfo.rs
@@ -7,7 +7,7 @@
 /// ```rust
 /// use bitcoin_rpc_codegen::client::latest::getmininginfo;
 ///
-/// let client = Client::new("http://127.0.0.1:8332", auth);
+/// let client = Client::new("http://127.0.0.1:18443", auth);
 /// let result = client.getmininginfo().await?;
 /// ```
 
@@ -19,12 +19,12 @@ use transport::{TransportTrait, TransportError};
 pub struct GetmininginfoResponse {
     pub blocks: u64,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub currentblockweight: Option<f64>,
+    pub currentblockweight: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub currentblocktx: Option<f64>,
+    pub currentblocktx: Option<u64>,
     pub difficulty: f64,
-    pub networkhashps: f64,
-    pub pooledtx: f64,
+    pub networkhashps: u64,
+    pub pooledtx: u64,
     pub chain: String,
     pub warnings: Vec<serde_json::Value>,
 }
